requirements.txt
aiogram
python-dotenv
openai
tiktoken